#217 내부 DB 쿼리시 에러 수정
